@@ -30,25 +30,17 @@ public final class TextProgressView extends LinearLayout {
 	private boolean mPostedShow = false;
 	private boolean mDismissed = false;
 
-	private final Runnable mDelayedHide = new Runnable() {
-
-		@Override
-		public void run() {
-			mPostedHide = false;
-			mStartTime = -1;
-			setVisibility(View.GONE);
-		}
+	private final Runnable mDelayedHide = () -> {
+		mPostedHide = false;
+		mStartTime = -1;
+		setVisibility(View.GONE);
 	};
 
-	private final Runnable mDelayedShow = new Runnable() {
-
-		@Override
-		public void run() {
-			mPostedShow = false;
-			if (!mDismissed) {
-				mStartTime = System.currentTimeMillis();
-				setVisibility(View.VISIBLE);
-			}
+	private final Runnable mDelayedShow = () -> {
+		mPostedShow = false;
+		if (!mDismissed) {
+			mStartTime = System.currentTimeMillis();
+			setVisibility(View.VISIBLE);
 		}
 	};
 
