@@ -25,7 +25,7 @@ public final class PagesCache {
 		}
 
 		private int compare(long x, long y) {
-			return (x < y) ? -1 : ((x == y) ? 0 : 1);
+			return Long.compare(x, y);
 		}
 	};
 
