@@ -7,7 +7,6 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.util.AttributeSet;
 import android.view.GestureDetector;
-import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewStub;
 import android.widget.FrameLayout;
@@ -78,13 +77,7 @@ public final class PageView extends FrameLayout implements View.OnClickListener,
 
 	@SuppressLint("ClickableViewAccessibility")
 	public void setTapDetector(final GestureDetector detector) {
-		mSubsamplingScaleImageView.setOnTouchListener(new OnTouchListener() {
-
-			@Override
-			public boolean onTouch(View v, MotionEvent event) {
-				return detector.onTouchEvent(event);
-			}
-		});
+		mSubsamplingScaleImageView.setOnTouchListener((v, event) -> detector.onTouchEvent(event));
 	}
 
 	public void setData(MangaPage page) {
